+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All Notable changes to `bakame/tokei` will be documented in this file.
 ### Added
 
 - Support for PHP8.6 `Time\Duration` using a Polyfill.
+- `Duration::seconds` and `Duration::microseconds`
 
 ### Fixed
 
@@ -19,7 +20,7 @@ All Notable changes to `bakame/tokei` will be documented in this file.
 
 ### Removed
 
-- None
+- **BC Break:**  `Duration` component public properties are removed use `Duration::$seconds` and `Duration::$microseconds` instead
 
 ## [0.3.0 - Chidorisō](https://github.com/bakame-php/stackwatch/compare/0.2.0...0.3.0) - 2026-07-21